requirements.txt
discord.py==1.6.0
numpy==1.16.2
pandas==0.24.2